--- src/components/frame/Extensions/Header/index.ts
@@ -1,2 +1,0 @@
-export { default as Header } from './Header';
-export { default as ColonyHeader } from './ColonyHeader';
