+++ src/components/frame/Extensions/layouts/index.ts
@@ -1,2 +1,4 @@
 export { default as MainLayout } from './MainLayout';
 export { default as ColonyLayout } from './ColonyLayout';
+export { default as Header } from './Header';
+export { default as ColonyHeader } from './ColonyHeader';
