+++ src/components/frame/Extensions/layouts/types.ts
@@ -1,9 +1,19 @@
-import { MessageDescriptor } from 'react-intl';
+import type { ReactNode } from 'react';
+import type { MessageDescriptor } from 'react-intl';
+
+import type { Colony } from '~types';
 
 export interface ColonyLayoutProps {
   title: MessageDescriptor;
   description: MessageDescriptor;
   loadingText: MessageDescriptor | string;
-  pageName: 'members' | 'extensions' | 'profile';
+  pageName: string;
   hideColonies?: boolean;
+}
+
+export interface HeaderProps {
+  colony?: Colony;
+  navBar?: ReactNode;
+  txButtons?: ReactNode;
+  userHub?: ReactNode;
 }
